TEXNO/requirements.txt
# MAX Bot
maxapi==0.1.0
python-dotenv==1.0.0

# Web Admin
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
aiofiles==23.2.1
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1